Fixed resource statistics for Docker containers being destroyed.
If a process has exited, but not reaped yet (zombie procses),
`/proc/<pid>/cgroup` will still exist, but the process's cgroup will be
reset to the root cgroup. In DockerContainerizer, we rely on
`/proc/<pid>/cgroup` to get the cpu/memory statistics of the container.
If the `usage` call happens when the process is a zombie, the cpu/memory
statistics will actually be that of the root cgroup, which is obviously
not correct. See more details in MESOS-8480.

This patch fixed the issue by checking if the cgroup of a given pid is
root cgroup or not.

Review: https://reviews.apache.org/r/65301/