Fixed connection refused error in IOSwitchboard for unix socket.
Once IOSwitchboard's unix socket appears in the file system, the agent
can try to connect to it. If the agent attempts to connect to the
socket before `listen()` has been called, then it gets connection
refused error. To address this, we initialize a unix socket using a
provisional path and rename it after `listen()` has been called.

Review: https://reviews.apache.org/r/65261/