Prevented redundant descriptions when applying reviewboard reviews.
Reviewboard's `rbt` can be configured to guess fields when creating
review requests. With that setting it e.g., uses the first line of a
commit as summary, and the remainder as description. If the commit
message has no body it will use the summary as description as well,
leading to redundant information in commit messages which committers
usually remove again manually before ultimately committing.

This patch prevents reusing a reviewboard description which is
identical to the summary when synthesising commit message with
`apply_reviews.py`.

Review: https://reviews.apache.org/r/65208/