Introduced `mesos-build` along with pre-built docker images.
The layout of `mesos-build` is similar to `mesos-tidy`.

The `mesos-build.sh` will be the user-facing driver, and `buildbot.sh`
invokes `mesos-build.sh` along with the clean-up (`docker rmi`) phase.
The `mesos-build` directory contains docker files and a common
`entrypoint.sh`. The docker images are built with all of the
dependencies required, and the various testing configurations are set
within `entrypoint.sh`.

Review: https://reviews.apache.org/r/65464