Updated `reservation.md`.
When using the Operator API, I came to this document to get a better
understand of resource reservations. It wasn't clear to me that
the reservation or reservations keys were a requirement given that
the principal configuration is optional. What I do know is that
the role can be set in the resource object or the reservation object.
I think we might as well instruct users to set the role in
the reservation object so it is more clear that it is required to
successfully POST the request. After a quick google search, I noted
that another user ran into the same issue and posted about it in IRC.

This closes #262