Windows: Fixed handle inheritance in `create_process` wrapper.
The primary change in this patch is that `create_process` now enables
inheritance for the `pipe` handles passed before starting the child
process. This is required, otherwise the child process will behave
incorrectly (for example, it will write to `stdout` but that will go
nowhere, as the redirection silently failed). After the process is
created, inheritance is disabled to prevent further calls to
`create_process` from inheriting the wrong handles.

The `std::tuple<os::WindowsFD, ...>` type was changed to a
`std::array<os::WindowsFD, 3>` as it is significantly easier to work
with (it supports iteration).

Review: https://reviews.apache.org/r/65574