Removed unnecessary compiler flag in build instructions.
After the recent upgrade to the bundled boost version, it
is not necessary anymore to add `-Wno-deprecated-declaration`
to the compiler flags.

Since users of Mesos shouldn't be expected to fix warnings in
our code, a reference to the `--disable-werror` configure flag
was also provided to allow easily ignoring compiler warnings.

Review: https://reviews.apache.org/r/65656/