Windows: Fixed CMake configuration for Ninja.
Only the Visual Studio generator(s) should check the toolset. The
Ninja generator requires the environment to be setup in advance, and
does not have the concept of a toolset, so this warning broke the
Ninja build.

Instead of warning that we're deprecating some arbitrary generator, we
should just say we don't officially support it so the user knows it
may not work as expected.

Fixed the assumption that all generators on Windows are
multi-configuration. Ninja is single-configuration, so the debug CRT
flag must sometimes be set for `CMAKE_CXX_FLAGS` depending on
`CMAKE_BUILD_TYPE`.

Review: https://reviews.apache.org/r/65719