Removed the unnecessary tmp dir during packaging docker build.
On Jenkins, we observed that tmp dirs are cleaned up quite frequently.
Since the build process will take about 30min, it's highly likely that
the tmp dir used in the script is cleaned up when we try to use it.

This patch removed the need for the tmp dir in the script.

Review: https://reviews.apache.org/r/65703