Reverted unintended edit to cpplint.py.
In 'bea8c3c9' we replaced '> >' with '>>' in C++ templates in a sweep
across the codebase. We also unintentionally modified a comment in
'cpplint.py' which was e.g., not reflected in its patch file and also
lead to no functional changes.

This patch reverts this modification to 'cpplint.py'.

Review: https://reviews.apache.org/r/64608/