Fixed the cgroup issue in the Mesos mini container.
For each cgroup subsystem, Docker does a bind mount from the current
cgroup to the root of the cgroup subsystem. For instance:
/sys/fs/cgroup/memory/docker/<cid> -> /sys/fs/cgroup/memory

This will confuse Mesos agent because it relies on proc file
'/proc/<pid>/cgroup` to determine the cgroup of a given process, and
this proc file is not affected by the bind mount.

This patch added a workaround to recreate the original cgroup
environment by doing another bind mount for each subsystem.