Added download button for master and agent logs in Web UI.
Adds a new button to download the logs in addition to the ability to
view them. Does not display the buttons if the master or agent does not
have an external log file or a log directory. Displays the correct logs,
i.e. if you are on a non-leading master you will get the logs of that
master and not the leading master.

Regarding the architectural choices, `HomeCtrl` is now similar to the
other controllers with an update function called when the state changes.

Review: https://reviews.apache.org/r/65246/