Windows: Ported `mesos-master` for testing.
This patch ports the `mesos-master` executable to Windows. Note that
this is only for testing, as we still lack support for a persistent
registry through `leveldb`, the master can only be run with
`--registry=in_memory` on Windows. However, this with the CLI is
enough to test the whole Mesos stack on a single Windows node.

The change was to temporarily remove support for the
`--ip_discovery_command` flag on Windows, since it uses a deleted
function, `os::shell`, instead emit an appropriate error message. When
`os::shell` is implemented on Windows, we will add support for
`--ip_discovery_command` back to the agent and master.

Review: https://reviews.apache.org/r/65799/