Windows: Fixed bug with CPU isolator not checking the max boundary.
The `set_job_cpu_limit` function does not sanitize its input, e.g.
`cpus`, causing a bug when more CPUs were requested than exist on the
machine.

For instance, in Mesos it was not possible to start a single task
which consumed every CPU, because the executor overcommit would cause
the task to ask for 0.1 more CPUs than available. So on a quad-core
machine, it'd ask for 4 CPUs, but that would cause the limiter to set
a limit at 4.1 CPUs, which exceeds the maximum allowed, causing an
"Invalid parameter" error when setting the CPU limit. We can avoid
this by ensuring the `cpu_rate` value is in the range of [1, 10000],
regardless of the input.

Review: https://reviews.apache.org/r/65891