Simplified system requirements of balloon executor.
The ballon executor demonstrates that if memory is isolated, executors
can get terminated should they exceed limits. For that the executor
allocates increasing amounts of memory. To make sure that the isolator
measurement is not skewed in the presence of swap, the executor
additionally locks the memory to make sure all allocated memory is
accounted for by the isolator.

Since the amount of memory a process can restricted on the system
level with `RLIMIT_MEMLOCK`, running this example can be impossible on
some setups.

This patch removes unneeded locking of memory on systems not using
swap. With that this example can be run on more setups, e.g., under
restricted accounts if no swap is used.

Review: https://reviews.apache.org/r/65921/