Windows: Removed signal macros in `checks/checker_process.cpp`.
The `__nestedCommandCheck()` implementation in `checker_process.cpp`
uses signals to determine if the task has been killed. This logic has
been removed on Windows, since signals aren't used on Windows. A
different implementation will be needed when nested command checks
work on Windows, but we remove it for now in order to build.

Review: https://reviews.apache.org/r/65862/