Windows: Fixed `WIFEXITED` and `WIFSIGNALED` stubs.
The `WIFEXITED` and `WIFSIGNALED` macros were incorrectly checking if
the exit code was not -1 to determine if the process exited or was
signaled. However, -1 is a valid return code on Windows, so when logic
like `CHECK(WIFEXITED(status)|| WIFSIGNALED(status))` was used, it
would end up aborting the process accidentally.

For `WIFEXITED`, we simply return `true` because all error codes on
Windows indicate the process exited (if the process instead failed to
spawn, then `os::spawn()` would return `None()` instead of an exit
code).

For `WIFIGNALED`, we simply return `false` for similar reasons. We
assume the process did not exit due to a signal, as that is not an
expected scenario on Windows.

Review: https://reviews.apache.org/r/65840/