Windows: Changed sharing semantics in `get_handle_follow`.
In the internal Windows functions `get_handle(_no)_follow`, we
erroneously only opened with `FILE_SHARE_READ` permissions. While we
are only opening the files for reading (indicated by the
`GENERIC_READ` flag), we need to open with write and delete sharing
semantics too, in order to avoid obtaining a lock on the file.

A regression test using `os::realpath` on a file that was already
being written to was added to ensure the correct behavior.

Review: https://reviews.apache.org/r/65936/