Fixed crash on macOS during certain authorization calls.
The method in charge of creating `ObjectApprovers` was relying that
iterating over equivalent hashset's would yield the same order of
results.

While that seemed to be true in different architectures, the macOS
implementation did not follow the same pattern showing a bug in the
code.

Review: https://reviews.apache.org/r/66088