Windows: Fixed the HTTP API path variables.
Many Mesos users assume that path separators are forward slashes, such
as dcos-ui, marathon-ui, and dcos-cli. Previous to this patch, if a
forward slash was given in the path variable to an HTTP API function
call to the Mesos agent on a Windows system, the Windows API would
fail at recognizing the path, because the Windows API accepts only
backslashes as path separators. To remedy this issue, we will now
convert all forward slashes passed as a path to the HTTP API to back
slashes on Windows agents by using the `path::from_uri` function.

Review: https://reviews.apache.org/r/66002/