Handled 'None' passed from the MasterDetector in 'Master::detect()'.
The function `MasterDetector::detect()` returns a value of type
`Future<Option<MasterInfo>>`, which, according to its documentation,
can be `None` if an election occured and no master is elected.

However, the code in `Master::detected()` was only handling the
cases of a failed future or a valid `MasterInfo` object.

*NOTE*: This commit does not add a corresponding unit test, since
that would require starting a non-leading master. For the
ZooKeeperMasterDetector, this is blocked by MESOS-2976, and an API
change to make this possible with the StandaloneMasterDetector
would add a lot of complexity to the `cluster::Master::start()`
function for a feature that is unlikely to be re-used in any other
test.

Review: https://reviews.apache.org/r/65571/