Removed SlaveTest.ChangeDomain.
This test was originally written with the assumption that the
would give get a new agent id after receiving a Shutdown
message from the master.

When that behaviour was changed, the test was disabled since
the agent could not reuse the previous agent id when a fault
domain was added.

With the introduction of reconfiguration policies, this feature
is now tested by SlaveTest.ReconfigurationPolicy, making this test
obsolete.

Review: https://reviews.apache.org/r/66248/