Changed the signature of a DefaultExecutor kill method.
This method is being passed the container to be killed as
`Owned<Container>`.

The container object is owned by the executor and stored
in `LinkedHashMap<TaskID, Owner<Container>> containers`.

We should pass around raw pointers and not make copies of
the stored `Owned` object, so this patch changes the
signature of the `kill` method.

Review: https://reviews.apache.org/r/65692/