Fixed an oversubscription test for agent registration backoff.
In the `OversubscriptionTest.ForwardUpdateSlaveMessage` test we
observe a single `UpdateSlaveMessage` to make sure the agent has fully
recovered. This message is sent from the resource provider-capable
agent to communicate its (empty) set of resource providers after
registration.

Since message was sent with a running clock, it is possible that the
agent encounters a timeout of its registration backoff timers. The
agent would then register agent, triggering another similar message
which is not expected in the test.

This patch adjusts the test to always run with paused clock
eliminating this particular scenario.

Review: https://reviews.apache.org/r/66313/