Revert "Fixed disconnection for ATTACH_CONTAINER_INPUT call...".
This reverts commit 02ebf9986ab5ce883a71df72e9e3392a3e37e40e.

Though this patch does fix the intended issue, it apparently surfaces
a bug in libprocess, which leads to a deadlock during IOSwitchBoard
finalization. Until the libprocess bug is fixed, it is better to
drop a single message and have one flaky tests than have a process
stuck during the cleanup and have multiple flaky tests and likely
issues in production.