Windows: Deleted `stout/os/windows/signals.hpp`.
This file had never been included before, so it didn't compile.
Because everything in this file will remain unimplemented, we chose to
delete the Windows version of the file entirely. The POSIX
implementation is left intact in `stout/os/posix/signals.hpp`, and the
primary header, `stout/os/signals.hpp` is a no-op on Windows. We do
this so that its inclusion does not need to be guarded, only the use
of the POSIX-specific APIs under `os::signals`.

Review: https://reviews.apache.org/r/66444