Fixed a potential race in `Sequence`.
Adding items to a sequence is accomplished by dispatching
`add()` to the sequence actor. However, this could
race with the sequence actor termination. Since the sequence
destructor discards the sequence's futures, and callers may
have registered callbacks which perform cleanup when the
futures are discarded, this race may mean that cleanup logic
fails to execute upon sequence destruction.

This patch fixes this by enqueueing the terminate
message at the end of the message queue.

Also removed the clock settle in the test `DiscardAll`,
as the processing of the messages is now guaranteed
to happen before actor termination.

Also added comments to clarify discard propagation.

Review: https://reviews.apache.org/r/66322/