Added a test to verify that task launch order is preserved.
The agent should launch tasks in the same order in which they are
received. In the task launch path, there are currently two
asynchronous steps which may complete out of order: unschedule GC
and task authorization.

This test simulates the reordering of the completion of the
unschedule GC step and verifies that, despite the reordering,
tasks will still launch in their original order.

Review: https://reviews.apache.org/r/66145/