Made sure test agent has reached stable state before starting test.
We require a certain number of offers in
`SlaveTest.ResourceProviderPublishAll`, so we need to make sure that
the agent knows all resources available during the test before offers
are being generated.

In this patch we await all `UpdateSlaveMessage` from the agent to make
sure the agent has seen all resources. Only then do we register a
framework. This greatly minimizes the risk of not seeing the expected
number of offers which now should only happen if due to
master-internal message reordering the allocator sees the framework
before it sees the latest agent resource update.

Review: https://reviews.apache.org/r/66508/