Windows: Changed `SLEEP_COMMAND` to use `ping`.
We have gone back to `ping` as the `sleep` command on Windows, because
PowerShell is too slow to startup, and can crash when the machine is
overloaded. This time, however, the output of `ping` is redirected to
`NUL` so that the tests aren't spammed with spurious output when
sleeping.

Review: https://reviews.apache.org/r/66415/