Fixed CLI bootstrap script to work with long workspace paths.
If the absolute path to the Python interpreter is long, it may exceed
the maximum length allowed for a shebang line (limit set to 128 on many
Linux distributions). When the shebang length limit is exceeded, pip
fails with the error: "bad interpreter: No such file or directory".

A work-around for this problem is to run the pip library module which is
what we do in this patch. This will also be done to use PyInstaller.

Review: https://reviews.apache.org/r/65705/