Used `csi::v0::VolumeCapability` in disk profile adaptors.
To adapt the change in CSI package names (from `csi` to `csi.v0`), we
now use `csi::v0::VolumeCapability` in the disk profile adaptor module
interface.

NOTE: This is not future-proof if there is a breaking change in
`VolumeCapability`, we might need to change the module interface to
support multiple CSI versions in the future.

Review: https://reviews.apache.org/r/66409/