Fixed Windows build by guarding Unix socket code.
We kept the `csi/paths.cpp` file in the build on Windows because
`slave/paths.cpp` now depends on it, and the only "broken" part is the
use of Unix sockets in `getEndpointSocketPath()`. For now, we will
return an error instead of trying to use unsupported Unix sockets on
Windows.