Added MemoryProfiler class to Libprocess.
This class exposes profiling functionality of jemalloc
memory allocator when it is detected to be the memory
allocator of the current process.

In particular, it enables developers to access live
statistics of current memory usage, and to create
heap profiles that show where most memory is allocated.

Review: https://reviews.apache.org/r/63368/