Changed call from `os::shell()` to `os::system()`.
The function `os::shell` is not implemented on Windows, making it
necessary to either hide all calls behind a preprocessor macro or to
avoid the function. Since we don't need the output of the child
process, we just switch to `os::system`.

Also, we included the missing header `<stout/os/write.hpp>`.

Review: https://reviews.apache.org/r/66700/