Made 3rdparty jemalloc only build when enabled.
When declaring a library with `add_library()`, the library is added to
the list of default targets, such that `make all` (and equivalents)
will build it, even if nothing depends on it. This meant that on
Windows, `cmake --build .` tried to run the `configure` script of
jemalloc, which obviously fails. By wrapping the 3rdparty import with
`ENABLE_JEMALLOC_ALLOCATOR` we fix this bug on Windows, and avoid
unnecessary build steps on Linux.

We do not use `EXCLUDE_FROM_ALL` so that we're consistent with the
rest of the CMake build (e.g. `libevent` is also wrapped with its
feature flag), and with Autotools (e.g. if the feature flag is not
set, it is not added at all).

Review: https://reviews.apache.org/r/66711