Fixed the ordering of Mesos containerizer isolators.
Historically, isolators were invoked in the order listed by
the operator in the `--isolation` flag. This changed to an
internal ordering in 6fb9c024, back to the operator ordering
in af474a6fa, and then to an undefined ordering in 9642d3c67b.

This commit switches back to an internal ordering for all the built-in
isolators. Custom isolators loaded in modules are run in operator order
after all the built-in ones. The rationale for an internal ordering is
expressed in MESOS-5581; basically we should not burden the operator
with having to figure out how to make the order correct. In the case of
custom isolators there's no way for us to know the correct ordering so
we make an arbitrary choice.

Review: https://reviews.apache.org/r/62472/