Windows: Fixed build with OpenSSL due to missing header.
When `-DENABLE_SSL=TRUE`, then `USE_SSL_SOCKET` is defined, and in
this `src/local/local.cpp` in a section guarded by said definition,
the construct `os::Permissions` is used. However, this is defined in
`stout/os/permissions.hpp`, which is implicitly included on POSIX, but
not on Windows. The fix is to IWYU and explicitly include it.

Review: https://reviews.apache.org/r/66753/