Removed `developers` from Java binding mesos.pom file.
This field should list information about whom to contact for feedback
on specific pieces of the released Java package, which largely
overlaps with the Apache committer concept.

The information in this file in general duplicates information already
in `docs/committers.md` which is rendered on the Mesos project
website.  We also publish a link to the Mesos developers mailing list
with the bindings which should be sufficient to reach Mesos
committers.

Review: https://reviews.apache.org/r/66762