Inserted explicit moves in libprocess.
In this patch we insert explicit moves in places where otherwise the
value would have been copied. In particular, an argument passed as an
rvalue will be available as an lvalue in a functions body, and an
explicit cast back to an rvalue is required to avoid creating a copy.
Similarly, when the type of the return value is different from the
type of the value being returned a temporary will be constructed to
convert to the correct type, and might be optimized due to RVO, but
nothing in (N)RVO would cause the conversion to use RVO; inserting an
explicit move will allow the compiler to select a potentially more
efficient way to construct the return value.

These instances were identified by a new move-related diagnostic added
recently to clang-7.

Review: https://reviews.apache.org/r/66779