Fixed incorrect enum value comparisons.
Due to ADL the enum values found originally found in this function where
from the v0 namespace while the value we compared to was from the v1
namespace. While this might work should the enum values in v0 and v1 map
to the same enum values, it does not seem to have been the intention and
is brittle.

This patch uses slightly longer names to refer to the enum values which
do not map onto symbols found via ADL.

Review: https://reviews.apache.org/r/66903/